feat: add firebase settings
