+++ firebase.json
@@ -1,0 +1,66 @@
+{
+	"flutter": {
+		"platforms": {
+			"android": {
+				"default": {
+					"projectId": "scenarioshelf-dev",
+					"appId": "1:700034805800:android:eed0450d58649e1bfa451d",
+					"fileOutput": "android/app/google-services.json"
+				},
+                "stg": {
+					"projectId": "scenarioshelf-stg",
+					"appId": "1:580165016681:android:2754e18bb031d5223556f8",
+					"fileOutput": "android/app/google-services.json"
+				},
+                "prod": {
+					"projectId": "scenarioshelf",
+					"appId": "1:710312653371:android:2ff28a53d327905c8f621e",
+					"fileOutput": "android/app/google-services.json"
+				}
+			},
+			"ios": {
+				"default": {
+					"projectId": "scenarioshelf-dev",
+					"appId": "1:700034805800:ios:c38d92d983ca06f5fa451d",
+					"uploadDebugSymbols": true,
+					"fileOutput": "ios/Runner/GoogleService-Info.plist"
+				},
+                "stg": {
+					"projectId": "scenarioshelf-stg",
+					"appId": "1:580165016681:ios:4fb857f7c07671413556f8",
+					"uploadDebugSymbols": true,
+					"fileOutput": "ios/Runner/GoogleService-Info.plist"
+				},
+                "prod": {
+					"projectId": "scenarioshelf",
+					"appId": "1:710312653371:ios:e14ca9ec05f6c8728f621e",
+					"uploadDebugSymbols": true,
+					"fileOutput": "ios/Runner/GoogleService-Info.plist"
+				}
+			},
+			"dart": {
+				"lib/repositories/firebase/firebase_options/dev/firebase_options.dart": {
+					"projectId": "scenarioshelf-dev",
+					"configurations": {
+						"android": "1:700034805800:android:eed0450d58649e1bfa451d",
+						"ios": "1:700034805800:ios:c38d92d983ca06f5fa451d"
+					}
+				},
+				"lib/repositories/firebase/firebase_options/stg/firebase_options.dart": {
+					"projectId": "scenarioshelf-stg",
+					"configurations": {
+						"android": "1:580165016681:android:2754e18bb031d5223556f8",
+						"ios": "1:580165016681:ios:4fb857f7c07671413556f8"
+					}
+				},
+				"lib/repositories/firebase/firebase_options/prod/firebase_options.dart": {
+					"projectId": "scenarioshelf",
+					"configurations": {
+						"android": "1:710312653371:android:2ff28a53d327905c8f621e",
+						"ios": "1:710312653371:ios:e14ca9ec05f6c8728f621e"
+					}
+				}
+			}
+		}
+	}
+}
