+++ ios/Flutter/Release.xcconfig
@@ -1,4 +1,3 @@
 #include? "Pods/Target Support Files/Pods-Runner/Pods-Runner.release.xcconfig"
 #include "Generated.xcconfig"
 #include "DartDefines.xcconfig"
-
