--- lib/repositories/firebase/firebase_options.dart
@@ -1,122 +1,0 @@
-// File generated by FlutterFire CLI.
-// ignore_for_file: lines_longer_than_80_chars, avoid_classes_with_only_static_members
-
-import 'package:flutter/foundation.dart' show TargetPlatform, defaultTargetPlatform, kIsWeb;
-
-import 'package:firebase_core/firebase_core.dart' show FirebaseOptions;
-
-import 'package:scenarioshelf/utils/environment.dart';
-
-/// Default [FirebaseOptions] for use with your Firebase apps.
-///
-/// Example:
-/// ```dart
-/// import 'firebase_options.dart';
-/// // ...
-/// await Firebase.initializeApp(
-///   options: DefaultFirebaseOptions.currentPlatform,
-/// );
-/// ```
-class DefaultFirebaseOptions {
-  static FirebaseOptions get currentPlatform {
-    if (kIsWeb) {
-      throw UnsupportedError(
-        'DefaultFirebaseOptions have not been configured for web - '
-        'you can reconfigure this by running the FlutterFire CLI again.',
-      );
-    }
-    switch (defaultTargetPlatform) {
-      case TargetPlatform.android:
-        switch (Environment.flavor) {
-          case Flavor.dev:
-            return androidDev;
-          case Flavor.stg:
-            return androidStg;
-          case Flavor.prod:
-            return androidProd;
-        }
-      case TargetPlatform.iOS:
-        switch (Environment.flavor) {
-          case Flavor.dev:
-            return iosDev;
-          case Flavor.stg:
-            return iosStg;
-          case Flavor.prod:
-            return iosProd;
-        }
-      case TargetPlatform.macOS:
-        throw UnsupportedError(
-          'DefaultFirebaseOptions have not been configured for macos - '
-          'you can reconfigure this by running the FlutterFire CLI again.',
-        );
-      case TargetPlatform.windows:
-        throw UnsupportedError(
-          'DefaultFirebaseOptions have not been configured for windows - '
-          'you can reconfigure this by running the FlutterFire CLI again.',
-        );
-      case TargetPlatform.linux:
-        throw UnsupportedError(
-          'DefaultFirebaseOptions have not been configured for linux - '
-          'you can reconfigure this by running the FlutterFire CLI again.',
-        );
-      case TargetPlatform.fuchsia:
-        throw UnsupportedError(
-          'DefaultFirebaseOptions are not supported for this platform.',
-        );
-    }
-  }
-
-  static const FirebaseOptions androidDev = FirebaseOptions(
-    apiKey: 'AIzaSyD3-T7-gwECyudvctTGqL-ZSJAd9duTg3I',
-    appId: '1:700034805800:android:eed0450d58649e1bfa451d',
-    messagingSenderId: '700034805800',
-    projectId: 'scenarioshelf-dev',
-    storageBucket: 'scenarioshelf-dev.appspot.com',
-  );
-
-  static const FirebaseOptions iosDev = FirebaseOptions(
-    apiKey: 'AIzaSyCPuXW4Hf-OiRNqkYA0ygcjljt-iXxS73Y',
-    appId: '1:700034805800:ios:c38d92d983ca06f5fa451d',
-    messagingSenderId: '700034805800',
-    projectId: 'scenarioshelf-dev',
-    storageBucket: 'scenarioshelf-dev.appspot.com',
-    iosClientId: '700034805800-6sa0qsspi2ue0m4brkv5mc9uc5nvmp9v.apps.googleusercontent.com',
-    iosBundleId: 'jp.scenarioshelf.dev',
-  );
-
-  static const FirebaseOptions androidStg = FirebaseOptions(
-    apiKey: 'AIzaSyDR89KbWPJ2Mdk2rk1pW7jZLLjZiMG_x3g',
-    appId: '1:580165016681:android:2754e18bb031d5223556f8',
-    messagingSenderId: '580165016681',
-    projectId: 'scenarioshelf-stg',
-    storageBucket: 'scenarioshelf-stg.appspot.com',
-  );
-
-  static const FirebaseOptions iosStg = FirebaseOptions(
-    apiKey: 'AIzaSyA9gf6NJnE8FTwOWIQ-rlZ-bCMCf1kUnDI',
-    appId: '1:580165016681:ios:4fb857f7c07671413556f8',
-    messagingSenderId: '580165016681',
-    projectId: 'scenarioshelf-stg',
-    storageBucket: 'scenarioshelf-stg.appspot.com',
-    iosClientId: '580165016681-rdbne9b6pcjne9p4m1r663gjrsso54i4.apps.googleusercontent.com',
-    iosBundleId: 'jp.scenarioshelf.stg',
-  );
-
-  static const FirebaseOptions androidProd = FirebaseOptions(
-    apiKey: 'AIzaSyD3XmbTaaZn9lB-IWq0vhHkpRhKV8rAb_Q',
-    appId: '1:710312653371:android:2ff28a53d327905c8f621e',
-    messagingSenderId: '710312653371',
-    projectId: 'scenarioshelf',
-    storageBucket: 'scenarioshelf.appspot.com',
-  );
-
-  static const FirebaseOptions iosProd = FirebaseOptions(
-    apiKey: 'AIzaSyDKqaFdG-_MAS58o1relgtPVuas_hgGzlc',
-    appId: '1:710312653371:ios:e14ca9ec05f6c8728f621e',
-    messagingSenderId: '710312653371',
-    projectId: 'scenarioshelf',
-    storageBucket: 'scenarioshelf.appspot.com',
-    iosClientId: '710312653371-svtul27sale65c46ejc15hotueikv0cj.apps.googleusercontent.com',
-    iosBundleId: 'jp.scenarioshelf',
-  );
-}
