+++ lib/repositories/firebase/firebase_options/prod/firebase_options.dart
@@ -1,0 +1,69 @@
+// File generated by FlutterFire CLI.
+// ignore_for_file: type=lint
+import 'package:firebase_core/firebase_core.dart' show FirebaseOptions;
+import 'package:flutter/foundation.dart'
+    show defaultTargetPlatform, kIsWeb, TargetPlatform;
+
+/// Default [FirebaseOptions] for use with your Firebase apps.
+///
+/// Example:
+/// ```dart
+/// import 'firebase_options.dart';
+/// // ...
+/// await Firebase.initializeApp(
+///   options: DefaultFirebaseOptions.currentPlatform,
+/// );
+/// ```
+class DefaultFirebaseOptions {
+  static FirebaseOptions get currentPlatform {
+    if (kIsWeb) {
+      throw UnsupportedError(
+        'DefaultFirebaseOptions have not been configured for web - '
+        'you can reconfigure this by running the FlutterFire CLI again.',
+      );
+    }
+    switch (defaultTargetPlatform) {
+      case TargetPlatform.android:
+        return android;
+      case TargetPlatform.iOS:
+        return ios;
+      case TargetPlatform.macOS:
+        throw UnsupportedError(
+          'DefaultFirebaseOptions have not been configured for macos - '
+          'you can reconfigure this by running the FlutterFire CLI again.',
+        );
+      case TargetPlatform.windows:
+        throw UnsupportedError(
+          'DefaultFirebaseOptions have not been configured for windows - '
+          'you can reconfigure this by running the FlutterFire CLI again.',
+        );
+      case TargetPlatform.linux:
+        throw UnsupportedError(
+          'DefaultFirebaseOptions have not been configured for linux - '
+          'you can reconfigure this by running the FlutterFire CLI again.',
+        );
+      default:
+        throw UnsupportedError(
+          'DefaultFirebaseOptions are not supported for this platform.',
+        );
+    }
+  }
+
+  static const FirebaseOptions android = FirebaseOptions(
+    apiKey: 'AIzaSyD3XmbTaaZn9lB-IWq0vhHkpRhKV8rAb_Q',
+    appId: '1:710312653371:android:2ff28a53d327905c8f621e',
+    messagingSenderId: '710312653371',
+    projectId: 'scenarioshelf',
+    storageBucket: 'scenarioshelf.appspot.com',
+  );
+
+  static const FirebaseOptions ios = FirebaseOptions(
+    apiKey: 'AIzaSyDKqaFdG-_MAS58o1relgtPVuas_hgGzlc',
+    appId: '1:710312653371:ios:e14ca9ec05f6c8728f621e',
+    messagingSenderId: '710312653371',
+    projectId: 'scenarioshelf',
+    storageBucket: 'scenarioshelf.appspot.com',
+    iosClientId: '710312653371-svtul27sale65c46ejc15hotueikv0cj.apps.googleusercontent.com',
+    iosBundleId: 'jp.scenarioshelf',
+  );
+}
